Move netmesh APIs into the pkg directory
This commit moves the netmesh protobuf API file and all generated code into
the pkg/nsm/apis/netmesh directory to match all of the existing Network
Service Mesh APIs.
