+++ plugins/objectstore/networkservice_test.go
@@ -17,8 +17,8 @@ package objectstore
 import (
 	"testing"
 
-	"github.com/ligato/networkservicemesh/netmesh/model/netmesh"
 	"github.com/ligato/networkservicemesh/pkg/nsm/apis/common"
+	"github.com/ligato/networkservicemesh/pkg/nsm/apis/netmesh"
 )
 
 type networkserviceStore struct {
