+++ plugins/objectstore/networkservicechannels.go
@@ -17,7 +17,7 @@ package objectstore
 import (
 	"sync"
 
-	"github.com/ligato/networkservicemesh/netmesh/model/netmesh"
+	"github.com/ligato/networkservicemesh/pkg/nsm/apis/netmesh"
 )
 
 // NetworkServiceChannelsStore map stores all discovered Network Service Channel
