+++ plugins/objectstore/networkserviceendpoints.go
@@ -17,7 +17,7 @@ package objectstore
 import (
 	"sync"
 
-	"github.com/ligato/networkservicemesh/netmesh/model/netmesh"
+	"github.com/ligato/networkservicemesh/pkg/nsm/apis/netmesh"
 )
 
 // NetworkServiceEndpointsStore map stores all discovered Network Service Endpoint
